Common: Game engine expects to know its current fps rate to do some calculation
Set the value back

+++ src/globals.h
@@ -89,7 +89,7 @@ extern uint32_t ff8;
 extern uint32_t ff8_currentdisk;
 
 extern uint32_t frame_counter;
-extern uint32_t frame_rate;
+extern double frame_rate;
 
 extern double speedhack_current;
